Refactors UI structure and removes unused logging
Modularizes the UI logic by extracting components into a dedicated `LinterUi` module, improving maintainability and readability. Removes redundant state management and unused imports in the main application file. Eliminates unnecessary console logging for cleaner output. Relocates and renames CSS for better organization.
